mv isHeader() to documentation helpers.
